// src/types.ts
export interface Game {
  "Game Title": string;
  Platform: string;
  Release: string;
  "1st Play": string;
  Gameplay: string;
  "World/Setting": string;
  Story: string;
  Replayability: string;
  Graphics: string;
  Sound: string;
  Nostalgia: string;
  Overall: string;
  Rating: string;
  Beaten: string;
  Completion: string;
  Comments: string;
}